+++ crane/version_basic.go
@@ -6,7 +6,7 @@ import (
 	"fmt"
 )
 
-const Version = "3.3.1"
+const Version = "3.3.2"
 const Pro = false
 
 func printVersion() {